Make default screen options into a reusable module
