+++ src/libs/Navigation/AppNavigator/defaultScreenOptions.js
@@ -1,0 +1,9 @@
+const defaultScreenOptions = {
+    cardStyle: {
+        overflow: 'visible',
+    },
+    headerShown: false,
+    animationTypeForReplace: 'pop',
+};
+
+export default defaultScreenOptions;
